% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tick.R
\name{tick}
\alias{tick}
\title{Timings for functions using mocking}
\usage{
tick(fn, pkg = NULL, ..., measure = "process")
}
\arguments{
\item{fn}{The un-namespaced name of the function to be timed, as a string.}

\item{pkg}{The name of the package where mocked functions should be inserted.}

\item{...}{Unused, should be empty.}

\item{measure}{One of \code{'process'} or \code{'real'}. Determines which of the
outputs of \code{bench::system_time()} (by the same names) will be logged, where
\code{'process'} is the process CPU usage of the expression evaluation and \code{'real'}
is the wallclock time.}
}
\description{
The stopwatch package introduces "tickers," which temporarily mock an
inputted function with a wrapper of the function that records the
elapsed time of the call to the inputted function and stores it for
later exploration.
}
\section{Workflow}{


The typical workflow of the package is as follows:
\itemize{
\item Establish a ticker for a function using \code{\link[=tick]{tick()}} like
\code{fn_enticked <- tick("fn", "pkg")}.
\item Run the code that will call \code{"fn"}.
\item Retrieve the timings for the call to \code{"fn"} with \code{\link[=ticks]{ticks()}}, like
\code{ticks(fn_enticked)}.
\item Restore the previous definition of \code{"fn"} with \code{\link[=untick]{untick()}}, like
\code{untick(fn_enticked)} (or \code{untick("fn", "pkg")}).
}
}

\examples{
# set up a "ticker" for `stats::lm()`:
lm_ticker <- tick("lm", "stats")
lm_ticker

# the function definition is temporarily modified:
stats::lm

# ...but it works as usual:
lm_res <- stats::lm(mpg ~ ., mtcars)
coef(lm_res)

# retrieve the timings associated with the ticker
ticks(lm_ticker)

# timings are logged for every call to the function
lm_res2 <- stats::lm(mpg ~ ., mtcars)
lm_res3 <- stats::lm(mpg ~ ., mtcars)

ticks(lm_ticker)

# restore the function to its original definition and clear timings
untick(lm_ticker)
stats::lm

}
\seealso{
\code{\link[=untick]{untick()}} and \code{\link[=ticks]{ticks()}}.
}
